Added Custom hash for unordered map
As unordered map works on hashing which is O(1) in average cases but sometimes problem setters design test cases in coding contests in such a manner that the normal unordered map goes in worst case i.e. O(N). So, this custom hash is to eliminate that chance. It runs in O(1) in all cases.